+++ weepickle/test/src/com/rallyhealth/weepickle/v1/FailureTests.scala
@@ -162,86 +162,5 @@ object FailureTests extends TestSuite {
       //      FromJson("10e-1").transform(ToScala[Int]) ==> 1
       //      FromJson("10e-1").transform(ToScala[Long]) ==> 1
     }
-    // format: off
-    test("tooManyFields"){
-      val b63 = Big63(
-        0, 1, 2, 3, 4, 5, 6, 7,
-        8, 9, 10, 11, 12, 13, 14,
-        15, 16, 17, 18, 19, 20, 21,
-        22, 23, 24, 25, 26, 27, 28,
-        29, 30, 31, 32, 33, 34, 35,
-        36, 37, 38, 39, 40, 41, 42,
-        43, 44, 45, 46, 47, 48, 49,
-        50, 51, 52, 53, 54, 55, 56,
-        57, 58, 59, 60, 61, 62
-      )
-      val b64 = Big64(
-        0, 1, 2, 3, 4, 5, 6, 7,
-        8, 9, 10, 11, 12, 13, 14,
-        15, 16, 17, 18, 19, 20, 21,
-        22, 23, 24, 25, 26, 27, 28,
-        29, 30, 31, 32, 33, 34, 35,
-        36, 37, 38, 39, 40, 41, 42,
-        43, 44, 45, 46, 47, 48, 49,
-        50, 51, 52, 53, 54, 55, 56,
-        57, 58, 59, 60, 61, 62, 63
-      )
-      val b65 = Big65(
-        0, 1, 2, 3, 4, 5, 6, 7,
-        8, 9, 10, 11, 12, 13, 14,
-        15, 16, 17, 18, 19, 20, 21,
-        22, 23, 24, 25, 26, 27, 28,
-        29, 30, 31, 32, 33, 34, 35,
-        36, 37, 38, 39, 40, 41, 42,
-        43, 44, 45, 46, 47, 48, 49,
-        50, 51, 52, 53, 54, 55, 56,
-        57, 58, 59, 60, 61, 62, 63,
-        64
-      )
-      // format: on
-      implicit val b63rw: com.rallyhealth.weepickle.v1.WeePickle.FromTo[Big63] =
-        com.rallyhealth.weepickle.v1.WeePickle.macroFromTo
-      implicit val b64rw: com.rallyhealth.weepickle.v1.WeePickle.FromTo[Big64] =
-        com.rallyhealth.weepickle.v1.WeePickle.macroFromTo
-      implicit val b65rw: com.rallyhealth.weepickle.v1.WeePickle.FromTo[Big65] =
-        com.rallyhealth.weepickle.v1.WeePickle.macroFromTo
-      val written63 = FromScala(b63).transform(ToJson.string)
-      assert(FromJson(written63).transform(ToScala[Big63]) == b63)
-      val written64 = FromScala(b64).transform(ToJson.string)
-      assert(FromJson(written64).transform(ToScala[Big64]) == b64)
-      val written65 = FromScala(b65).transform(ToJson.string)
-      assert(FromJson(written65).transform(ToScala[Big65]) == b65)
-    }
   }
 }
-
-// format: off
-case class Big63(_0: Byte, _1: Byte, _2: Byte, _3: Byte, _4: Byte, _5: Byte, _6: Byte, _7: Byte,
-  _8: Byte, _9: Byte, _10: Byte, _11: Byte, _12: Byte, _13: Byte, _14: Byte,
-  _15: Byte, _16: Byte, _17: Byte, _18: Byte, _19: Byte, _20: Byte, _21: Byte,
-  _22: Byte, _23: Byte, _24: Byte, _25: Byte, _26: Byte, _27: Byte, _28: Byte,
-  _29: Byte, _30: Byte, _31: Byte, _32: Byte, _33: Byte, _34: Byte, _35: Byte,
-  _36: Byte, _37: Byte, _38: Byte, _39: Byte, _40: Byte, _41: Byte, _42: Byte,
-  _43: Byte, _44: Byte, _45: Byte, _46: Byte, _47: Byte, _48: Byte, _49: Byte,
-  _50: Byte, _51: Byte, _52: Byte, _53: Byte, _54: Byte, _55: Byte, _56: Byte,
-  _57: Byte, _58: Byte, _59: Byte, _60: Byte, _61: Byte, _62: Byte)
-case class Big64(_0: Byte, _1: Byte, _2: Byte, _3: Byte, _4: Byte, _5: Byte, _6: Byte, _7: Byte,
-  _8: Byte, _9: Byte, _10: Byte, _11: Byte, _12: Byte, _13: Byte, _14: Byte,
-  _15: Byte, _16: Byte, _17: Byte, _18: Byte, _19: Byte, _20: Byte, _21: Byte,
-  _22: Byte, _23: Byte, _24: Byte, _25: Byte, _26: Byte, _27: Byte, _28: Byte,
-  _29: Byte, _30: Byte, _31: Byte, _32: Byte, _33: Byte, _34: Byte, _35: Byte,
-  _36: Byte, _37: Byte, _38: Byte, _39: Byte, _40: Byte, _41: Byte, _42: Byte,
-  _43: Byte, _44: Byte, _45: Byte, _46: Byte, _47: Byte, _48: Byte, _49: Byte,
-  _50: Byte, _51: Byte, _52: Byte, _53: Byte, _54: Byte, _55: Byte, _56: Byte,
-  _57: Byte, _58: Byte, _59: Byte, _60: Byte, _61: Byte, _62: Byte, _63: Byte)
-case class Big65(_0: Byte, _1: Byte, _2: Byte, _3: Byte, _4: Byte, _5: Byte, _6: Byte, _7: Byte,
-  _8: Byte, _9: Byte, _10: Byte, _11: Byte, _12: Byte, _13: Byte, _14: Byte,
-  _15: Byte, _16: Byte, _17: Byte, _18: Byte, _19: Byte, _20: Byte, _21: Byte,
-  _22: Byte, _23: Byte, _24: Byte, _25: Byte, _26: Byte, _27: Byte, _28: Byte,
-  _29: Byte, _30: Byte, _31: Byte, _32: Byte, _33: Byte, _34: Byte, _35: Byte,
-  _36: Byte, _37: Byte, _38: Byte, _39: Byte, _40: Byte, _41: Byte, _42: Byte,
-  _43: Byte, _44: Byte, _45: Byte, _46: Byte, _47: Byte, _48: Byte, _49: Byte,
-  _50: Byte, _51: Byte, _52: Byte, _53: Byte, _54: Byte, _55: Byte, _56: Byte,
-  _57: Byte, _58: Byte, _59: Byte, _60: Byte, _61: Byte, _62: Byte, _63: Byte,
-  _64: Byte)
-// format: on
